Flyout Focus trap and refactor
