+++ designer/client/components/Flyout/index.ts
@@ -1,0 +1,1 @@
+export { Flyout } from "./Flyout";
