+++ designer/package.json
@@ -41,6 +41,7 @@
     "core-js": "^3.8.1",
     "dotenv": "8.2.0",
     "flat": "5.0.0",
+    "focus-trap-react": "^8.3.2",
     "hapi-pino": "^8.0.1",
     "hoek": "^6.1.3",
     "i18next": "^19.7.0",
